Update FileCreator.sh
Converted the prompt method into a menu where the user can choose between creating a file or removing it.

Made the structure of the FileCreatePrompt more seamless using if statements to check if the users chosen location exists (if not, they will be prompted to create the directory.) or if they wish for the file to be created onto the desktop.